Design and implement features for JcClasspath #57
