+++ jacodb-core/src/main/kotlin/org/jacodb/impl/bytecode/JcClassOrInterfaceImpl.kt
@@ -102,11 +102,36 @@ class JcClassOrInterfaceImpl(
         }
 
     override val declaredFields: List<JcField> by lazy(LazyThreadSafetyMode.NONE) {
-        info.fields.map { JcFieldImpl(this, it) }
+        val fields = info.fields
+        val result: List<JcField> = fields.map { JcFieldImpl(this, it) }
+        when {
+            !features.isNullOrEmpty() -> {
+                val modifiedFields = result.toMutableList()
+                features.forEach {
+                    it.fieldsOf(this)?.let {
+                        modifiedFields.addAll(it)
+                    }
+                }
+                modifiedFields
+            }
+            else -> result
+        }
     }
 
     override val declaredMethods: List<JcMethod> by lazy(LazyThreadSafetyMode.NONE) {
-        info.methods.map { toJcMethod(it, classSource, features) }
+        val result: List<JcMethod> = info.methods.map { toJcMethod(it, classSource, features) }
+        when {
+            !features.isNullOrEmpty() -> {
+                val modifiedMethods = result.toMutableList()
+                features.forEach {
+                    it.methodsOf(this)?.let {
+                        modifiedMethods.addAll(it)
+                    }
+                }
+                modifiedMethods
+            }
+            else -> result
+        }
     }
 
     override fun equals(other: Any?): Boolean {
